+++ src/components/TodoItem/index.ts
@@ -1,2 +1,1 @@
-// eslint-disable-next-line
 export * from './TodoItem';
